Make autoDecaySeconds a double instead of int

+++ server/map.h
@@ -206,7 +206,7 @@ char *getMapChangeLineString( MapChangeRecord *inRecord,
 // returns number of seconds from now until when next decay is supposed
 // to happen
 // returns -1 if no decay pending
-int getNextDecayDelta();
+double getNextDecayDelta();
 
 
 // marks region as looked at, so that live decay tracking continues
